+++ src/Resources/contao/languages/en/default.php
@@ -44,6 +44,7 @@ $GLOBALS['TL_LANG']['PA2']['imageSortTypes']['custom'] = array('Own sorting', 'O
  */
 $GLOBALS['TL_LANG']['PA2']['albumPreviewImageTypes']['no_preview_image'] = array('No preview image', 'No preview image');
 $GLOBALS['TL_LANG']['PA2']['albumPreviewImageTypes']['random_preview_image'] = array('Random preview image', 'Random preview image');
+$GLOBALS['TL_LANG']['PA2']['albumPreviewImageTypes']['first_preview_image'] = array('First preview image', 'First preview image');
 $GLOBALS['TL_LANG']['PA2']['albumPreviewImageTypes']['select_preview_image'] = array('Select preview image', 'Select preview image');
 
 /**
@@ -52,7 +53,6 @@ $GLOBALS['TL_LANG']['PA2']['albumPreviewImageTypes']['select_preview_image'] = a
 $GLOBALS['TL_LANG']['PA2']['previewImageModuleTypes']['use_album_options'] = array('Settings of album take over', 'Settings of album take over');
 $GLOBALS['TL_LANG']['PA2']['previewImageModuleTypes']['no_preview_images'] = array('No preview images', 'No preview images');
 $GLOBALS['TL_LANG']['PA2']['previewImageModuleTypes']['random_images'] = array('Random preview images', 'Random preview images');
-$GLOBALS['TL_LANG']['PA2']['previewImageModuleTypes']['first_image'] = array('First preview image', 'First preview image');
 $GLOBALS['TL_LANG']['PA2']['previewImageModuleTypes']['random_images_at_no_preview_images'] = array('Random preview images at an undefined use preview image', 'Random preview images at an undefined use preview image');
 $GLOBALS['TL_LANG']['PA2']['previewImageModuleTypes']['first_image_at_no_preview_images'] = array('First preview image at an undefined use preview image', 'First preview image at an undefined use preview image');